bugfixes
consertado bug em que Cliente.formatar_cpf() levantava um erro mesmo em cpf válido.
consertado bug de importação circular em Clientes.py e Cliente.py (Cliente.py foi excluido, agora as duas classes estão contidas em Clientes.py)
Resolvido bug onde __salvo da classe Clientes não ia se tornar False quando deveria
removido novo_cliente da classe Clientes, agora no método add_cliente o objeto Cliente deve ser fornecido externamente
